+++ test/fixtures/ts-module-minification/svelte5.svelte
@@ -1,0 +1,9 @@
+<script lang="ts">
+  import { a1 } from "./module.svelte.ts";
+
+  let test = a1;
+</script>
+
+<button>
+  {test}
+</button>